add comments for the clean code
