+++ lib/ui/widgets/add_task_sheet.dart
@@ -4,11 +4,14 @@ import 'package:todo_list/models/task_model.dart';
 import 'package:todo_list/services/task/task_service.dart';
 import 'package:todo_list/ui/pages/viewmodels/add_task_view_model.dart';
 
+// Bottom sheet widget used to create a new task.
+// This widget provides an AddTaskViewModel to manage form state.
 class AddTaskSheet extends StatelessWidget {
   const AddTaskSheet({super.key});
 
   @override
   Widget build(BuildContext context) {
+    // Inject ViewModel into widget tree
     return ChangeNotifierProvider(
       create: (_) => AddTaskViewModel(),
       child: const _AddTaskSheetBody(),
@@ -18,24 +21,30 @@ class AddTaskSheet extends StatelessWidget {
 
 class _AddTaskSheetBody extends StatelessWidget {
   const _AddTaskSheetBody();
+
   @override
   Widget build(BuildContext context) {
+    // ViewModel holds all form logic (title, tags, category, image …)
     final vm = context.watch<AddTaskViewModel>();
+
     return Padding(
       padding: EdgeInsets.only(
         left: 16,
         right: 16,
         top: 16,
-        bottom: MediaQuery.of(context).viewInsets.bottom + 16,
+        bottom: MediaQuery.of(context).viewInsets.bottom + 16, // handle keyboard
       ),
       child: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
           const _SheetHandle(),
           const SizedBox(height: 8),
+
+          // Title section
           Text('New task', style: Theme.of(context).textTheme.titleLarge),
           const SizedBox(height: 12),
 
+          // TITLE INPUT
           TextField(
             controller: vm.titleCtrl,
             decoration: const InputDecoration(
@@ -46,6 +55,7 @@ class _AddTaskSheetBody extends StatelessWidget {
           ),
           const SizedBox(height: 12),
 
+          // DESCRIPTION INPUT
           TextField(
             controller: vm.descCtrl,
             decoration: const InputDecoration(
@@ -55,10 +65,13 @@ class _AddTaskSheetBody extends StatelessWidget {
             ),
             maxLines: 3,
           ),
+
           const SizedBox(height: 16),
 
+          // CATEGORY SELECTION
           Text('Category', style: Theme.of(context).textTheme.titleSmall),
           const SizedBox(height: 8),
+
           Wrap(
             spacing: 8,
             runSpacing: 8,
@@ -68,6 +81,7 @@ class _AddTaskSheetBody extends StatelessWidget {
                     label: Row(
                       mainAxisSize: MainAxisSize.min,
                       children: [
+                        // Small color dot matching category color
                         Container(
                           width: 12,
                           height: 12,
@@ -86,12 +100,8 @@ class _AddTaskSheetBody extends StatelessWidget {
                     selectedColor: category.color.withValues(alpha: 0.3),
                     checkmarkColor: category.color,
                     labelStyle: TextStyle(
-                      color: vm.selectedCategory == category
-                          ? category.color
-                          : null,
-                      fontWeight: vm.selectedCategory == category
-                          ? FontWeight.w600
-                          : null,
+                      color: vm.selectedCategory == category ? category.color : null,
+                      fontWeight: vm.selectedCategory == category ? FontWeight.w600 : null,
                     ),
                   ),
                 )
@@ -100,11 +110,13 @@ class _AddTaskSheetBody extends StatelessWidget {
 
           const SizedBox(height: 16),
 
+          // TAGS SECTION
           Text('Tags', style: Theme.of(context).textTheme.titleSmall),
           const SizedBox(height: 8),
 
           Row(
             children: [
+              // Tag text field
               Expanded(
                 child: TextField(
                   controller: vm.tagCtrl,
@@ -117,18 +129,21 @@ class _AddTaskSheetBody extends StatelessWidget {
                   onSubmitted: vm.addTag,
                 ),
               ),
+
               const SizedBox(width: 8),
+
+              // Add tag button
               IconButton(
                 onPressed: () => vm.addTag(vm.tagCtrl.text),
                 icon: const Icon(Icons.add_circle),
                 style: IconButton.styleFrom(
-                  backgroundColor: Theme.of(
-                    context,
-                  ).colorScheme.primaryContainer,
+                  backgroundColor: Theme.of(context).colorScheme.primaryContainer,
                 ),
               ),
             ],
           ),
+
+          // Display added tags
           if (vm.tags.isNotEmpty) ...[
             const SizedBox(height: 8),
             Wrap(
@@ -140,9 +155,8 @@ class _AddTaskSheetBody extends StatelessWidget {
                       label: Text('#$tag'),
                       onDeleted: () => vm.removeTag(tag),
                       deleteIcon: const Icon(Icons.close, size: 16),
-                      backgroundColor: Theme.of(
-                        context,
-                      ).colorScheme.surfaceContainerHighest,
+                      backgroundColor:
+                          Theme.of(context).colorScheme.surfaceContainerHighest,
                     ),
                   )
                   .toList(),
@@ -150,9 +164,12 @@ class _AddTaskSheetBody extends StatelessWidget {
           ],
 
           const SizedBox(height: 20),
+
+          // PHOTO SECTION
           Text('Photo', style: Theme.of(context).textTheme.titleSmall),
           const SizedBox(height: 8),
 
+          // If no image selected
           if (vm.imageBytes == null)
             OutlinedButton.icon(
               onPressed: vm.pickImage,
@@ -160,6 +177,7 @@ class _AddTaskSheetBody extends StatelessWidget {
               label: const Text('Add a photo'),
             )
           else
+            // Image preview
             Column(
               children: [
                 ClipRRect(
@@ -172,6 +190,8 @@ class _AddTaskSheetBody extends StatelessWidget {
                   ),
                 ),
                 const SizedBox(height: 8),
+
+                // Buttons to remove or change photo
                 Row(
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: [
@@ -190,23 +210,30 @@ class _AddTaskSheetBody extends StatelessWidget {
                 ),
               ],
             ),
+
           const SizedBox(height: 16),
 
+          // ACTION BUTTONS
           Row(
             children: [
+              // Cancel button
               Expanded(
                 child: FilledButton.tonal(
                   onPressed: () => Navigator.pop(context),
                   child: const Text('Cancel'),
                 ),
               ),
+
               const SizedBox(width: 12),
+
+              // Add task button
               Expanded(
                 child: FilledButton(
                   onPressed: () async {
                     final service = context.read<TaskService>();
                     final title = vm.titleCtrl.text.trim();
 
+                    // Title required
                     if (title.isEmpty) {
                       ScaffoldMessenger.of(context).showSnackBar(
                         const SnackBar(content: Text('Please enter a title')),
@@ -214,9 +241,10 @@ class _AddTaskSheetBody extends StatelessWidget {
                       return;
                     }
 
+                    // Create task and upload image if provided
                     await service.addTask(
                       TaskModel(
-                        id: '_',
+                        id: '_', // id assigned by Firestore
                         userId: '_',
                         title: title,
                         description: vm.descCtrl.text.trim(),
@@ -230,7 +258,10 @@ class _AddTaskSheetBody extends StatelessWidget {
                     );
 
                     if (!context.mounted) return;
+
                     Navigator.pop(context);
+
+                    // Confirmation Snackbar
                     ScaffoldMessenger.of(context).showSnackBar(
                       SnackBar(
                         content: Text('Task "$title" added'),
@@ -249,8 +280,10 @@ class _AddTaskSheetBody extends StatelessWidget {
   }
 }
 
+// Small handle displayed at the top of the bottom sheet
 class _SheetHandle extends StatelessWidget {
   const _SheetHandle();
+
   @override
   Widget build(BuildContext context) {
     return Container(
